pyproject.toml
[project]
name = "ttd-bot"
version = "1.2.6"
description = "ttd-bot"
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "gunicorn>=23.0.0",
    "uvicorn[standard]>=0.35.0",
    "nonebot2[httpx,fastapi,aiohttp]>=2.4.3",
    "nb-cli>=1.4.2",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-orm[default]>=0.8.2",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot-plugin-localstore>=0.7.4",
    "nonebot-plugin-uninfo>=0.9.0",
    "nonebot-plugin-status>=0.9.0",
    "nonebot-plugin-chatrecorder>=0.7.0",
    "nonebot-plugin-wordcloud>=0.9.0",
    "nonebot-plugin-datastore>=1.3.1",
    "nonebot-plugin-add-friends>=0.1.5",
    "nonebot-plugin-auto-sendlike>=0.0.12",
    "nonebot-plugin-authrespond>=1.8",
    "nonebot-plugin-deer-pipe @ git+https://github.com/trytodupe/nonebot-plugin-deer-pipe@978b6ece2455f72c4c4a4ddc4763c129bd3c241a",
    "nonebot-plugin-learning-chat @ git+https://github.com/trytodupe/nonebot-plugin-learning-chat@45c39f2af8b79d2113a1fdc96e0a51de65427f56",
    "nonebot-plugin-auto-enter-group",
    "nonebot-plugin-analysis-bilibili>=2.8.1",
    "nonebot-plugin-mc-server-status>=0.4.1",
    "nonebot-plugin-clovers>=0.3.5",
    "nonebot-plugin-groupmate-waifu>=0.27",
    "clovers-setu-collection>=0.2.3.post1",
    "nonebot-plugin-essence-message @ git+https://github.com/trytodupe/nonebot-plugin-essence-message@d120b28fa38e5eff5516ea204831314bf136c256",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = [
    "nonebot_plugin_orm",
    "nonebot_plugin_localstore",
    "nonebot_plugin_uninfo",
    "nonebot_plugin_status",
    "nonebot_plugin_chatrecorder",
    "nonebot_plugin_clovers",
    "nonebot_plugin_wordcloud",
    "nonebot_plugin_datastore",
    "nonebot_plugin_add_friends",
    "nonebot-plugin-auto-sendlike",
    "nonebot_plugin_authrespond",
    "nonebot_plugin_deer_pipe",
    "nonebot_plugin_learning_chat",
    "nonebot_plugin_auto_enter_group",
    "nonebot_plugin_analysis_bilibili",
    "nonebot_plugin_groupmate_waifu",
    "nonebot_plugin_mc_server_status",
    "nonebot_plugin_essence_message",
]
plugin_dirs = ["src/plugins"]
builtin_plugins = ["echo"]
